pyproject.toml
[project]
name = "fastapi-url-shortener"
version = "0.1.0"
description = "Url script shortener"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "jinja2>=3.1.6",
    "pydantic-settings>=2.12.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "redis[hiredis]>=6.0.0",
    "typer>=0.15.2",
    "uvicorn<=0.20.0",
]

# uv settings
[tool.uv]
required-version = "~=0.9.7"

[dependency-groups]
dev = [
    "coverage>=7.9.2",
    "httpx>=0.28.1",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.3.0",
    "ruff>=0.11.12",
]

# mypy settings
[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]

# ruff settings
[tool.ruff]
required-version = ">=0.11.12"
# Exclude a variety of commonly ignored directories.
# Allow imports relative to the "src" and "test" directories.
src = ["url-shortener", ]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    ".venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.13
target-version = "py313"

[tool.ruff.lint]
select = [
    "FAST", # FastAPI (FAST)
    "ANN", # flake8-annotations (ANN)
    # "ASYNC", # flake8-async (ASYNC)
    "I", # isort (I)
    "F", # Pyflakes (F)
    "S", # flake8-bandit (S)
    "BLE", # flake8-blind-except (BLE)
    "FBT", # flake8-boolean-trap (FBT)
    "B", # flake8-bugbear (B)
    "COM", # flake8-commas (COM)
    "C4", # flake8-comprehensions (C4)
    "DTZ", # flake8-datetinez (DTZ)
    "T10", # flake8-debugger (T10)
    "EM", # flake8-errmsg (EM)
    "ISC", #flake8-implict-str-concat (ISC)
    "ICN", # flake8-import-conventions (ICN)
    "LOG", # flake8-logging (LOG)
    "G", # flake8-logging-format (G)
    "INP", # flake8-no-pep420 (INP)
    "PIE", # flake8-pie (PIE)
    "T20", # flake8-print (T20)
    "PYI", # flake8-pyi (PYI)
    #    "PT", # flake8-pytest-style (PT)
    "Q", # flake8-quotes (Q)
    "RSE", # flake8-raise (RSE)
    "RET", # flake8-return (RET)
    "SLF", # flake8-self(SLF)
    "SIM", # flake8-simplify (SIM)
    "SLOT", # flake8 -slots (SLOT)
    "TC", # flake8-type-checking (TCH)
    "ARG", # flake8-unused-arguments (ARG)
    "PTH", # flake8-use-pathlib (PTH)
    "FLY", # flynt (FLY)
    "C90", # mccabe (C90)
    "PD", # pandas-vet (PD)
    "N", # pep8-naming (N)
    "PERF", # Perflint (PERF)
    "E", # pycodestyle Error (E)
    "W", # pycodestyle (W)
    #    "DOC", # pydocrint (DOC)
    #     "D", # pycodestyle (D)
    "F", # Pyflakes (F)
    "PGH", # pygrep-hooks
    "PLC", # pylint Convention
    "PLE", # pylint Error
    "PLR", # pylint refactor
    "PLW", # pylint warning
    "UP", #pyupgrade
    "FURB", # refurb (FURB)
    "TRY", # tryceratops (TRY)
    "RUF", # Ruff-specific rules

]
ignore = [
    "S101", # S101 assert Use of assert detected
    "RUF001", # unicode string
    "RUF002", # unicode docstring
    "RUF003", # unicode comment
    "TC006", # runtime-cast-value
]

[tool.ruff.lint.per-file-ignores]
"url-shortener/testing/**/*.py" = ["S311"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []


[tool.pytest.ini_options]
minversion = 8.3 # минимальаня версия pytest
addopts = "-ra --strict-markers"
# -ra Отображается сводка по пропущенным/упавшим тестам
# --strict-markers Строгая проверка названий меток

log_cli = false # Отключите вывод логов в консоль
log_cli_level = "INFO" # Установите уровень логирования
console_output_style = "progress" # стиль вывода в консоль

# регистрация меток "название:описание"
markers = [
    "apitest:test any http call",
    "templatetest:test any http HTML template page"
]
xfail_strict = true # если в тесте не указано, то это значение по умолчанию
